% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_df.R
\name{convert_df}
\alias{convert_df}
\title{Convert to SimDesign Object}
\usage{
convert_df(x)
}
\arguments{
\item{x}{An \code{R} dataframe object, specifically with simulation design
variables specified as factors, and simulation results as numeric.}
}
\value{
An \code{R} dataframe object, of class SimDesign and data.frame.
}
\description{
\code{convert_df} is a helper function that
can convert a dataframe into a SimDesign object.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{SimDisplay}}
}

