% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_levels.R
\name{get_sim_levels}
\alias{get_sim_levels}
\title{Extract simulation levels}
\usage{
get_sim_levels(x)
}
\arguments{
\item{x}{A \code{data.frame} object, specifically of class \code{SimDesign}.}
}
\value{
A vector of type character.
}
\description{
\code{get_sim_levels} is a helper function that
returns the vector names of the design levels from a \code{SimDesign} object.
}
\examples{
\dontrun{
#TODO
}

}
\seealso{
\code{\link{SimDisplay}, \link{get_design_levels}}
}

