% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_suggests.R
\name{install_suggests}
\alias{install_suggests}
\title{Check for and install supplementary packages}
\usage{
install_suggests()
}
\description{
\code{install_suggests} is a helper function to be used when first
installing \code{SimDisplay}. As some of the procedures within this package
rely on functions from other packages, this function will check to ensure
all required libraries are available.
}
\examples{
\dontrun{
install_suggests()
}

}
\seealso{
\code{\link{SimDisplay}}
}
