% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_sim.R
\name{round_sim}
\alias{round_sim}
\title{Rounding values for presentation}
\usage{
round_sim(x, digits = 3)
}
\arguments{
\item{x}{A \code{vector} or \code{data.frame} object.}

\item{digits}{An \code{integer} value indicating the number of digits to round to.}
}
\value{
A \code{vector} or \code{data.frame} object.
}
\description{
\code{round_sim} takes a vector or data.frame object and returns the same
type of object, but rounded
}
\examples{
\dontrun{
data(Brown1974)
round_sim(Brown1974) # returns error
round_sim(Brown1974[,5:9]) # dataframe input, dataframe output
str(round_sim(Brown1974[,5:9])) # vectors are now character-type
round_sim(Brown1974[,5:9], 2)

dat <- rnorm(n = 50, mean = 0, sd = 150) # Wider range, vector input
round_sim(dat) # vector output
}

}
\seealso{
\code{\link{SimDisplay}}
}
