% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundSim.R
\name{roundSim}
\alias{roundSim}
\title{Rounding values for presentation}
\usage{
roundSim(x, digits = 3, percent = FALSE)
}
\arguments{
\item{x}{A \code{vector} or \code{data.frame} object.}

\item{digits}{An \code{integer} value indicating the number of digits to round to.}

\item{percent}{A \code{boolean} value indicating whether column elements should be multiplied by 100.}
}
\value{
A \code{vector} or \code{data.frame} object.
}
\description{
\code{roundSim} takes a vector or data.frame object and returns the same
type of object, but rounded.
}
\examples{
\dontrun{
data(Brown1974)
roundSim(Brown1974) # returns error
roundSim(Brown1974[,4:9]) # dataframe input, dataframe output
str(roundSim(Brown1974[,4:9])) # vectors are now character-type
roundSim(Brown1974[,4:9], 2)

set.seed(10)
dat <- rnorm(n = 5, mean = 0, sd = 150) # Wider range, vector input
roundSim(dat) # vector output
}

}
\seealso{
\code{\link{SimDisplay}}
}
