% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySim.R
\name{shinySim}
\alias{shinySim}
\title{Launch Shiny interface for MCSS results}
\usage{
shinySim(x)
}
\arguments{
\item{x}{An \code{R} dataframe object, specifically of class \code{SimDesign}.}
}
\description{
\code{shinySim}launches an interactive shiny interface
specifically designed for the analysis of simulation study results.
}
\examples{
\dontrun{
data(tsimresults)
str(tsimresults)
head(tsimresults)
library(shiny)
runApp(shinySim(tsimresults))
}

}
\seealso{
\code{\link{SimDisplay}}
}

