% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_emph.R
\name{sim_emph}
\alias{sim_emph}
\title{Emphasize cells from simulation results}
\usage{
sim_emph(dat, upper.bound = 0.075, lower.bound = 0.025, colnames = NULL,
  digits = 2)
}
\arguments{
\item{dat}{A \code{data.frame} object of \code{class(SimDesign)}.}

\item{upper.bound}{An \code{integer} value indicating the upper limit.}

\item{lower.bound}{An \code{integer} value indicating the lower limit.}

\item{colnames}{An optional \code{character} vector indicating column names to use in the output.}

\item{digits}{Display how many significant digits?}
}
\value{
A character \code{vector}.
}
\description{
\code{sim_emph} takes a \code{SimDesign} dataframe object, and returns a
}
\examples{
\dontrun{
data(Brown1974)
Brown1974$REPLICATIONS <- Brown1974$SIM_TIME <- Brown1974$COMPLETED <- Brown1974$SEED <- NULL
TypeI <- subset(Brown1974, var_ratio == 1) # Separate type I error rate conditions
Power <- subset(Brown1974, var_ratio != 1) # from power conditions
sim_emph(TypeI)
}

}
\seealso{
\code{\link{SimDisplay}}
}
