% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDisplay-package.R
\docType{package}
\name{SimDisplay-package}
\alias{SimDisplay-package}
\alias{SimDisplay}
\title{SimDisplay - Methods for visualizing complex tables.}
\description{
The SimDisplay package provides a variety of methods for visualizing
results from complex tables. It has been specifically designed to
summarise Monte Carlo Simulation Study objects from \code{\link{SimDesign}} (Chalmers, 2016).
}
\section{SimDisplay functions}{

So far, this package implements the following functions: \code{tableplot}.
}

\author{
Matthew Sigal \email{matthewsigal@gmail.com}
}
\keyword{package}
