% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyDf.R
\name{simplifyDf}
\alias{simplifyDf}
\title{Simplify SimDesign dataframe}
\usage{
simplifyDf(x)
}
\arguments{
\item{x}{A \code{SimDesign} \code{data.frame} object.}
}
\value{
A simplified \code{SimDesign} \code{data.frame} object.
}
\description{
\code{simplifyDf} takes a data.frame object of class \code{SimDesign} and
returns a data.frame without all of the additional attributes.
}
\examples{
\dontrun{
data(Brown1974)
str(Brown1974) # Note the many additional attributes
df <- simplifyDf(Brown1974)
str(df)
}

}
\seealso{
\code{\link{SimDisplay}}
}
