% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTableplot.R
\name{simTableplot}
\alias{simTableplot}
\title{Generate a tableplot.}
\usage{
simTableplot(dat, design_vars = NULL, main_title = "Tableplot")
}
\arguments{
\item{dat}{A \code{data.frame} object of \code{class(SimDesign)}.}

\item{design_vars}{An optional \code{character} vector pertaining to the design variables used to label each row. If omitted, all design variables will be used based upon the design levels attribute of the SimDesign object.}

\item{main_title}{A \code{character} value for the main title of the plot.}
}
\description{
\code{simTableplot} takes a \code{SimDesign} dataframe object, and returns a tableplot.
}
\examples{
\dontrun{
data(Brown1974)
TypeI <- subset(Brown1974, var_ratio == 1)
simTableplot(TypeI)
simTableplot(TypeI,
             design_vars = c("distribution", "sample_size"),
             main_title = "Type I error rates by distribution and sample size.")
}

}
\seealso{
\code{\link{SimDisplay}}
}
