% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline.R
\name{sparkline}
\alias{sparkline}
\title{Generate sparkline graphic}
\usage{
sparkline(x)
}
\arguments{
\item{x}{A \code{vector} object, pertaining to a row from a tabular display.}
}
\value{
A \code{ggplot2} graphic object.
}
\description{
\code{sparkline} takes a vector of type numeric and generates a simple sparkline graphic.
}
\examples{
\dontrun{
data(Brown1974)
}

}
\seealso{
\code{\link{SimDisplay}}
}
