% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableShade.R
\name{tableShade}
\alias{tableShade}
\title{Generate a shaded table.}
\usage{
tableShade(dat, table_vars = NULL, method = "mean", ndigit = 2,
  colswitch = 0.6, numrow = 2, colours = "Blues",
  main_title = "Shaded Table", xlab = NULL, ylab = NULL)
}
\arguments{
\item{dat}{A \code{data.frame} object of \code{class(SimDesign)}.}

\item{table_vars}{A \code{character} vector pertaining to the variables that will be presented on the display. The first element specifies the facetting variable, the second the X-axis.}

\item{method}{A \code{character} vector of length 1 pertaining to the method to be used to summarize results.}

\item{ndigit}{An \code{integer} indicating the number of decimal places to use in the labelling.}

\item{colswitch}{A \code{numeric} value indicating where to switch from black text to white.}

\item{numrow}{An \code{integer} value indicating the number of rows to use in the facet grid.}

\item{main_title}{A \code{character} value for the main title of the plot.}

\item{xlab}{A \code{character} value for the x-axis label.}

\item{ylab}{A \code{character} value for the y-axis label.}

\item{cols}{A \code{character} value indicating the palette to use.}
}
\value{
A \code{ggplot2} object.
}
\description{
\code{tableShade} takes a \code{SimDesign} dataframe object, and returns a shaded table.
}
\examples{
\dontrun{
data(Brown1974)
dat <- subset(Brown1974, var_ratio != 1)
tableShade(dat, table_vars = c("distribution", "var_ratio"))

# Adding optional arguments:
tableShade(dat, table_vars = c("distribution", "var_ratio"),
           main_title = "Power Rates by Variance Ratio and Distribution",
           xlab = "Variance Ratio", ylab = "Outcome Variable",
           numrow = 1, ndigit = 1, colswitch = .7, cols = "Reds")
}

}
\seealso{
\code{\link{SimDisplay}}
}
