% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDisplay-package.R
\docType{data}
\name{tsimresults}
\alias{tsimresults}
\title{Sample Simulation Results}
\format{A data frame with 36 rows and 7 variables}
\usage{
data(tsimresults)
}
\description{
\code{tsimresults} is a a dataset (of class data.frame as well as SimDesign) containing
the standard output from \code{SimDesign}.
For details on how this simulation was run, please refer to \code{?SimDesign::runSimulation}.
The variables in the dataframe are:
}
\details{
\itemize{
  \item sample_size. The sample size conditions in the study (30, 60, 90, 120)
  \item group_size_ratio. The difference in sample size between the groups (.5, 1, 2)
  \item standard_deviation_ratio. The difference in standard deviation between the groups (1, 4, 8)
  \item lessthan.05.welch. The number of cells with p-values less than .05, using the Welch correction.
  \item lessthan.05.independent. The number of cells with p-values less than .05, not using the Welch correction.
  \item REPLICATIONS. The number of replications for each cell.
  \item SIM_TIME. The amount of time each cell took to run.
}
}
\keyword{datasets}

