% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_df.R
\name{convert_df}
\alias{convert_df}
\title{Convert to SimDesign Object}
\usage{
convert_df(x)
}
\arguments{
\item{x}{A \code{data.frame} object, specifically with simulation design
variables specified as factors, and simulation results as numeric.}
}
\value{
A \code{data.frame} object also of class \code{SimDesign}.
}
\description{
\code{convert_df} is a helper function that
can convert a \code{data.frame} into a \code{SimDesign} object.
}
\examples{
\dontrun{
n <- c(15, 30, 50)
sd <- c(0.5, 1, 2)
skew <- c(1, 5, 10)
dat <- expand.grid(N = n, SD = sd, Skew = skew,
                   KEEP.OUT.ATTRS = FALSE)
set.seed(123)
dat$Result1 <- rnorm(27)
dat$Result2 <- rnorm(27)

convert_df(dat) # Error: N, SD, and Skew not set as factors

dat[,1:3] <- lapply(dat[,1:3], as.factor)
newdat <- convert_df(dat)
str(newdat)
}

}
\seealso{
\code{\link{SimDisplay}}
}
