% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design_levels.R
\name{get_design_levels}
\alias{get_design_levels}
\title{Extract design levels}
\usage{
get_design_levels(x)
}
\arguments{
\item{x}{An \code{R} dataframe object, specifically of class \code{SimDesign}.}
}
\value{
A vector of type character.
}
\description{
\code{get_design_levels} is a helper function that
returns the vector names of the design levels from a SimDesign object.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{SimDisplay}, \link{get_sim_levels}}
}

