% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySim.R
\name{shinySim}
\alias{shinySim}
\title{Launch Shiny interface for MCSS results}
\usage{
shinySim(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} object, specifically of class \code{SimDesign}.}

\item{...}{additional arguments to be passed to \code{\link{runApp}}}
}
\description{
\code{shinySim} launches an interactive shiny interface
specifically designed for the analysis of simulation study results.
}
\examples{
\dontrun{
data(tsimresults)
str(tsimresults)
head(tsimresults)

shinySim(tsimresults, launch.browser = TRUE)

}

}
\seealso{
\code{\link{SimDisplay}}
}

