% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySim.R
\name{shinySim}
\alias{shinySim}
\title{shinySim
\code{shinySim} is a function that launches an interactive shiny interface
specifically designed for the analysis of simulation study designs.}
\usage{
shinySim(x)
}
\arguments{
\item{x}{An \code{R} dataframe object, specifically of class \code{SimDesign}.}
}
\description{
shinySim
\code{shinySim} is a function that launches an interactive shiny interface
specifically designed for the analysis of simulation study designs.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{SimDisplay}}
}

