% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_design_check.R
\name{sim_design_check}
\alias{sim_design_check}
\title{sim_design_check}
\usage{
sim_design_check(x)
}
\arguments{
\item{x}{A \code{data.frame} object, specifically of class \code{SimDesign}.}
}
\value{
A \code{data.frame} object.
}
\description{
The \code{sim_design_check} function checks an R object to see if it is
of class \code{SimDesign}. If so, the design and simulation elements are
extracted and the design variables are converted to factors.
}
\examples{
\dontrun{
#TODO
}

}
\seealso{
\code{\link{SimDisplay}}
}

