% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTable.R
\name{simTable}
\alias{simTable}
\title{Emphasize cells from simulation results}
\usage{
simTable(dat, by = NULL, highlight = TRUE, upper.bound = 0.075,
  lower.bound = 0.025, colnames = NULL, lucid = TRUE, digits = 2,
  caption = NULL, hide = NULL)
}
\arguments{
\item{dat}{A \code{data.frame} object of \code{class(SimDesign)}.}

\item{by}{A \code{character} value indicating a factor variable to collapse results by.}

\item{highlight}{A \code{boolean} value indicating if values should be highlighted based upon \code{upper.bound} and \code{lower.bound}.}

\item{upper.bound}{An \code{integer} value indicating the upper limit.}

\item{lower.bound}{An \code{integer} value indicating the lower limit.}

\item{colnames}{An optional \code{character} vector indicating column names to use in the output.}

\item{lucid}{Should results be printed? This converts results to characters of same length.}

\item{digits}{Display lucid results to how many significant digits?}

\item{caption}{A \code{character} string indicating caption.}

\item{hide}{A \code{character} vector indicating the names of any columns to hide from the table.}
}
\value{
A character \code{vector}.
}
\description{
\code{simTable} takes a \code{SimDesign} dataframe object, and returns a LaTeX table.
}
\examples{
\dontrun{
data(Brown1974)
Brown1974$REPLICATIONS <- Brown1974$SIM_TIME <- Brown1974$COMPLETED <- Brown1974$SEED <- NULL
TypeI <- subset(Brown1974, var_ratio == 1) # Separate type I error rate conditions
Power <- subset(Brown1974, var_ratio != 1) # from power conditions
simTable(TypeI)
simTable(TypeI, by = 'sample_size')
simTable(TypeI, by = 'sample_size', hide = 'var_ratio')
}

}
\seealso{
\code{\link{SimDisplay}}
}
