% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTablePlot.R
\name{simTablePlot}
\alias{simTablePlot}
\title{\code{simTablePlot} extends \code{ggplot2} to produce table plot displays.}
\usage{
simTablePlot(x)
}
\arguments{
\item{x}{A \code{data.frame} object, specifically of class \code{SimDesign}.}
}
\description{
\code{simTablePlot} extends \code{ggplot2} to produce table plot displays.
}
\examples{
\dontrun{
#TODO
}

}
\seealso{
\code{\link{SimDisplay}}, \code{\link{runSimulation}}
}

