% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTablePlot.R
\name{simTablePlot}
\alias{simTablePlot}
\title{\code{simTablePlot} extends ggplot2 to produce table plot displays.}
\usage{
simTablePlot(x)
}
\arguments{
\item{x}{An \code{R} dataframe object, specifically of class \code{SimDesign}.}
}
\description{
\code{simTablePlot} extends ggplot2 to produce table plot displays.
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{SimDisplay}}
}

