% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline.R
\name{sparkline}
\alias{sparkline}
\title{Generate sparkline graphic}
\usage{
sparkline(x, min = 0, max = 1)
}
\arguments{
\item{x}{A \code{vector} object, pertaining to a row from a tabular display.}

\item{min}{Numeric input, specifying the minimum value for the y-axis}

\item{max}{Numeric input, specifying the maximum value for the y-axis}
}
\value{
A \code{ggplot2} graphic object.
}
\description{
\code{sparkline} takes a vector of type numeric and generates a simple sparkline graphic.
}
\examples{
\dontrun{
data(Brown1974)
sparkline(Brown1974[1,4:9])
}

}
\seealso{
\code{\link{SimDisplay}}
}
