% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.SimDesign}
\alias{summary.SimDesign}
\title{Summary method for SimDesign objects}
\usage{
\method{summary}{SimDesign}(x)
}
\arguments{
\item{x}{A \code{data.frame} object, of class \code{SimDesign}.}
}
\description{
\code{summary.SimDesign} is a default method for summarizing a
\code{data.frame} of class \code{SimDesign}.
}
\examples{
\dontrun{
summary(Brown1974)
type1 <- subset(Brown1974, var_ratio == 1)
power <- subset(Brown1974, var_ratio != 1)

summary(type1)
summary(power)
}

}
\seealso{
\code{\link{SimDisplay}}
}
